requirements.txt
cachetools==3.0.0
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-nested-admin==3.1.3
django-nested-inline==0.3.7
django-registration-redux==2.5
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
html5lib==1.0.1
httplib2==0.12.0
idna==2.7
lxml==4.2.5
oauth2client==4.1.3
oauthlib==2.1.0
Pillow==5.3.0
pipenv==2018.7.1
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyJWT==1.6.4
PyPDF2==1.26.0
python-docx==0.8.7
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.9
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
webencodings==0.5.1
whitenoise==4.0
xhtml2pdf==0.2.3
